refactor(tui): remove backwards-compat shims, use direct imports
- Delete diff_tui.py, view_tui.py, tui_base.py, catalog_tui.py shims
- Update CLI commands to import from tui/ package directly
- Update tests to import from tui/ package directly

Not released yet, so no backwards compatibility needed.

Authored by: Aaron Lippold<lippold@gmail.com>
