+++ tests/unit/test_catalog_browser.py
@@ -8,27 +8,27 @@ class TestCatalogBrowserAppExists:
 
     def test_catalog_browser_app_importable(self):
         """CatalogBrowserApp should be importable."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         assert CatalogBrowserApp is not None
 
     def test_catalog_browser_app_extends_base_browser(self):
         """CatalogBrowserApp should extend BaseBrowserApp."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
-        from cis_bench.cli.commands.tui_base import BaseBrowserApp
+        from cis_bench.cli.commands.tui import BaseBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         assert issubclass(CatalogBrowserApp, BaseBrowserApp)
 
     def test_catalog_detail_view_exists(self):
         """CatalogDetailView should be importable."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         assert CatalogDetailView is not None
 
     def test_catalog_detail_view_extends_detail_view(self):
         """CatalogDetailView should extend DetailView."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
-        from cis_bench.cli.commands.tui_base import DetailView
+        from cis_bench.cli.commands.tui import DetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         assert issubclass(CatalogDetailView, DetailView)
 
@@ -38,35 +38,35 @@ class TestCatalogBrowserBindings:
 
     def test_has_search_binding(self):
         """Catalog browser should have search binding (/)."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         binding_keys = [b.key for b in CatalogBrowserApp.BINDINGS]
         assert "slash" in binding_keys
 
     def test_has_space_for_multiselect(self):
         """Catalog browser should have space binding for multi-select."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         binding_keys = [b.key for b in CatalogBrowserApp.BINDINGS]
         assert "space" in binding_keys
 
     def test_has_jump_binding(self):
         """Catalog browser should have g binding for jump."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         binding_keys = [b.key for b in CatalogBrowserApp.BINDINGS]
         assert "g" in binding_keys
 
     def test_has_help_binding(self):
         """Catalog browser should have ? binding for help."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         binding_keys = [b.key for b in CatalogBrowserApp.BINDINGS]
         assert "question_mark" in binding_keys
 
     def test_has_quit_binding(self):
         """Catalog browser should have q binding for quit."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         binding_keys = [b.key for b in CatalogBrowserApp.BINDINGS]
         assert "q" in binding_keys
@@ -77,14 +77,14 @@ class TestCatalogBrowserActions:
 
     def test_has_toggle_select_action(self):
         """CatalogBrowserApp should have action_toggle_select method."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         assert hasattr(CatalogBrowserApp, "action_toggle_select")
         assert callable(CatalogBrowserApp.action_toggle_select)
 
     def test_has_apply_search_filter(self):
         """CatalogBrowserApp should have _apply_search_filter method."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         assert hasattr(CatalogBrowserApp, "_apply_search_filter")
         assert callable(CatalogBrowserApp._apply_search_filter)
@@ -121,7 +121,7 @@ class TestCatalogBrowserInitialization:
 
     def test_accepts_benchmarks_parameter(self, sample_benchmarks):
         """CatalogBrowserApp should accept benchmarks in constructor."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         app = CatalogBrowserApp(benchmarks=sample_benchmarks)
         assert hasattr(app, "_benchmarks")
@@ -129,7 +129,7 @@ class TestCatalogBrowserInitialization:
 
     def test_initializes_empty_selection(self, sample_benchmarks):
         """CatalogBrowserApp should initialize with empty selection set."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         app = CatalogBrowserApp(benchmarks=sample_benchmarks)
         assert hasattr(app, "_selected_indices")
@@ -138,7 +138,7 @@ class TestCatalogBrowserInitialization:
 
     def test_stores_all_benchmarks(self, sample_benchmarks):
         """CatalogBrowserApp should store all benchmarks for filtering."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         app = CatalogBrowserApp(benchmarks=sample_benchmarks)
         assert hasattr(app, "_all_benchmarks")
@@ -166,7 +166,7 @@ class TestCatalogDetailViewContent:
 
     def test_update_content_accepts_benchmark(self, sample_benchmark):
         """CatalogDetailView should have update_content method accepting benchmark dict."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         view = CatalogDetailView()
         # Should not raise
@@ -174,7 +174,7 @@ class TestCatalogDetailViewContent:
 
     def test_update_content_sets_text(self, sample_benchmark):
         """CatalogDetailView.update_content should set content text."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         view = CatalogDetailView()
         view.update_content(sample_benchmark)
@@ -183,7 +183,7 @@ class TestCatalogDetailViewContent:
 
     def test_content_includes_version(self, sample_benchmark):
         """Detail content should include benchmark version."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         view = CatalogDetailView()
         view.update_content(sample_benchmark)
@@ -192,7 +192,7 @@ class TestCatalogDetailViewContent:
 
     def test_content_includes_platform(self, sample_benchmark):
         """Detail content should include platform."""
-        from cis_bench.cli.commands.catalog_tui import CatalogDetailView
+        from cis_bench.cli.commands.tui.catalog import CatalogDetailView
 
         view = CatalogDetailView()
         view.update_content(sample_benchmark)
@@ -214,7 +214,7 @@ class TestCatalogBrowserMultiSelect:
 
     def test_get_selected_items_returns_list(self, sample_benchmarks):
         """get_selected_items should return a list."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         app = CatalogBrowserApp(benchmarks=sample_benchmarks)
         result = app.get_selected_items()
@@ -222,7 +222,7 @@ class TestCatalogBrowserMultiSelect:
 
     def test_get_selected_items_empty_initially(self, sample_benchmarks):
         """get_selected_items should return empty list initially."""
-        from cis_bench.cli.commands.catalog_tui import CatalogBrowserApp
+        from cis_bench.cli.commands.tui.catalog import CatalogBrowserApp
 
         app = CatalogBrowserApp(benchmarks=sample_benchmarks)
         result = app.get_selected_items()
@@ -234,6 +234,6 @@ class TestRunCatalogBrowserFunction:
 
     def test_run_catalog_browser_exists(self):
         """run_catalog_browser function should exist."""
-        from cis_bench.cli.commands.catalog_tui import run_catalog_browser
+        from cis_bench.cli.commands.tui.catalog import run_catalog_browser
 
         assert callable(run_catalog_browser)
